refactor: extract GITHUB API operatinos into a shared helper
